Refactor simulator start process and implement firmware watch
This commit includes improvements to the start process for the simulator along with implementation of firmware watch. It also renames files and classes to improve code maintainability. Now the WokwiSimulator class implements a new Simulator interface to allow the management of simulator processes more appropriately. The FirmwareWatcher helps to react to changes for any files that are being watched. In WokwiProjectService, the argsLoader service is introduced to handle input reading during a simulation start. Also, code clean-ups and revisions are addressed.