Refactor firmware format detection and command start logic
The implementation of determining the firmware format has been moved to a new utility method in FirmwareUtils. This refactor simplifies the code in WokwiArgsLoader. Also, an adjustment has been made to the start command in WokwiSimulator, appropriating the firmware format. A new string extension function hexStringToByteArray() for hexadecimal string conversion is also added.