Refactor simulator start logic in WokwiIntellij
This commit refactors the logic for starting the wokwi simulator in WokwiIntellij. The focus of the changes was to improve the reusability and efficiency of the code by extracting duplicated logic into a `startSimulatorRunExecution` method and replacing `startSimulatorSuspended` with `startSimulatorAsync`. Additionally, redundant and unused code was removed to clean up the overall structure.